Fix CONFERENCE table reference to CONFERENCE_ROOM
I had incorrectly added a one to one relationship between the tables, many to
one is the correct one. A one to one relationship ensures that a single
conference can only reference single room and vice versa. Many to one ensures
that a single conference can only have a single room but a single room can be
referenced by many conferences.

This took me way to long to realize.